refactor: postContainer
