+++ packages/gdsc-dju-blog/src/components/atoms/CategoryMenu/styled.ts
@@ -36,9 +36,10 @@ export const CategoryTextWrapper = styled(motion.div)<{ isActive: boolean }>`
 `;
 export const CategoryMenuWrapper = styled(motion.div)`
   display: flex;
-  align-items: center;
+  align-items: flex-end;
   flex-direction: row;
   width: 713px;
+  height: 70px;
   justify-content: space-between;
 `;
 export const CategoryCircleWrapper = styled(motion.div)`
