+++ packages/gdsc-dju-blog/src/components/atoms/Notice.tsx
@@ -1,21 +1,27 @@
 import React, { ReactNode } from 'react';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 
-export const NoticeText = styled.p`
+export const NoticeText = styled.p<{ align: boolean }>`
   display: flex;
   align-items: center;
-  justify-content: center;
-  height: 100%;
+  justify-content: flex-start;
+  width: 100%;
   font-size: ${({ theme }) => theme.fontSize.h6};
   color: ${({ theme }) => theme.colors.grey400};
+  ${({ align }) =>
+    align &&
+    css`
+      justify-content: center;
+    `}
 `;
 
 interface Props {
   children: React.ReactNode;
+  align?: boolean;
 }
 
-const Notice = ({ children }: Props) => {
-  return <NoticeText>{children}</NoticeText>;
+const Notice = ({ children, align = true }: Props) => {
+  return <NoticeText align={align}>{children}</NoticeText>;
 };
 
 export default Notice;
