+++ packages/gdsc-dju-blog/src/components/layouts/PostSavesLayout/styled.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const PostSavesLayoutContainer = styled.div`
+  display: flex;
+  gap: 50px;
+  flex-wrap: wrap;
+  margin-top: 100px;
+`;
