+++ packages/gdsc-dju-blog/src/components/molecules/PostCard/styled.ts
@@ -1,89 +1,143 @@
-import { motion } from 'framer-motion';
 import styled, { css } from 'styled-components';
+import { motion } from 'framer-motion';
 
-export const PostCardWrapper = styled(motion.div)`
-  display: flex;
-  align-items: center;
-  flex-direction: row;
+export const PostCardInner = styled(motion.article)`
   position: relative;
-  width: 100%;
-  cursor: pointer;
+  width: 248px;
+  height: 294px;
   border-radius: 10px;
-  box-shadow: ${({ theme }) => theme.colors.boxShadow100};
-  box-sizing: border-box;
-`;
-export const PostCardImage = styled.img`
-  //background: black;
-  height: 150%;
-  position: relative;
-`;
-export const PostCardImageWrapper = styled.div`
-  height: 100%;
-  width: 50%;
-  overflow: hidden;
   display: flex;
-  position: absolute;
-  align-items: center;
-  justify-content: flex-start;
-  border-bottom-left-radius: 10px;
-  border-top-left-radius: 10px;
+  flex-direction: column;
+  box-shadow: 0px 19px 32px -1px rgba(0, 0, 0, 0.1);
+  cursor: pointer;
 `;
 
-export const PostCardContentWrapper = styled(motion.div)<{ hover: boolean }>`
-  box-sizing: border-box;
-  height: 240px;
-  border-top-right-radius: 10px;
-  border-bottom-right-radius: 10px;
-  position: relative;
-  padding: 30px 40px;
-  z-index: 2;
-  background: ${({ theme }) => theme.colors.background};
-  left: 30%;
-  width: 70%;
-  flex-direction: column;
-  transition: all 0.3s ease-in-out;
-  ${({ hover }) =>
-    hover &&
-    css`
-      left: 5%;
-      width: 95%;
-    `}
-`;
-export const PostDate = styled.p`
-  font-size: ${({ theme }) => theme.fontSize.body2};
-  color: ${({ theme }) => theme.colors.grey600};
-  margin-bottom: 12px;
-`;
-export const BookmarkWrapper = styled.div`
+export const BookMarkWrapper = styled.div`
   position: absolute;
-  top: -4px;
-  right: 40px;
-  transition: all 0.3s ease;
-  z-index: 3;
+  width: 23.33px;
+  height: 30px;
+  right: 17px;
+  top: -3px;
+  cursor: pointer;
 `;
-export const PostTitle = styled.h4`
-  font-size: ${({ theme }) => theme.fontSize.h5};
+
+export const PostCardThumbnailWrapper = styled.div`
+  width: 248px;
+  height: 294px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 13px;
+  overflow: hidden;
+`;
+export const PostCardThumbnail = styled.img`
+  height: 294px;
+`;
+
+export const PostCardTagWrapper = styled.div<{ IsHovered: boolean }>`
+  display: flex;
+  position: absolute;
+  bottom: ${({ IsHovered }) => (IsHovered ? '254px' : '92px')};
+  transition: all 0.3s ease;
+  left: 9px;
+`;
+export const PostCardTag = styled.div`
+  height: 17px;
+  padding: 1px 10px;
+  border-radius: 50px;
+  border: 1px solid #ffffff;
+  box-sizing: border-box;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  span {
+    font-size: 10px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 15px;
+    text-align: left;
+    color: #ffffff;
+  }
+`;
+export const PostCardBottomBox = styled(motion.div)<{ isHovered: boolean }>`
+  position: absolute;
+  border-radius: 10px;
+  width: 100%;
+  bottom: 0;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  background: ${({ theme }) => theme.colors.cardBackground};
+  padding: 8px 20px 12px;
+  gap: ${({ isHovered }) => (isHovered ? '170px' : '8px')};
+  transition: all 0.3s ease;
+  cursor: pointer;
+  overflow: hidden;
+`;
+export const PostCardTitle = styled(motion.h6)`
+  font-size: ${({ theme }) => theme.fontSize.h6};
+  font-weight: 400;
   color: ${({ theme }) => theme.colors.grey900};
-  margin-bottom: 16px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+`;
+export const PostCardSubTextWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`;
+export const PostCardAuthorWrapper = styled.div`
+  display: flex;
+  align-items: center;
+`;
+export const PostCardAuthorImage = styled.img`
+  background: ${({ theme }) => theme.colors.grey400};
+  border-radius: 100%;
+  margin-right: 7px;
+  height: 18px;
+  width: 18px;
+`;
+export const PostCardPostText = styled(motion.div)`
+  position: absolute;
+  transform: translate(-50%, 0%);
+  top: 53px;
+
+  min-height: 120px;
+  max-height: 120px;
   word-break: break-word;
   overflow: hidden;
   text-overflow: ellipsis;
 `;
-export const PostHashTageSection = styled.section`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-bottom: 20px;
-  transition: all 0.2s;
-`;
-export const PostContent = styled(motion.p)`
-  font-size: ${({ theme }) => theme.fontSize.body1};
-  color: ${({ theme }) => theme.colors.grey600};
-  box-sizing: border-box;
-  transition: all 0.3s ease;
-  overflow: hidden;
+export const PostText = styled.article`
+  width: 208px;
+  white-space: normal;
   display: -webkit-box;
   -webkit-box-orient: vertical;
-  white-space: normal;
-  -webkit-line-clamp: 3;
+  overflow: hidden;
+  -webkit-line-clamp: 7;
+  font-size: ${({ theme }) => theme.fontSize.body3};
+  color: ${({ theme }) => theme.colors.grey700};
+`;
+
+export const PostCardSubText = styled.div<{
+  subText?: boolean;
+  bold?: boolean;
+}>`
+  font-size: ${({ theme }) => theme.fontSize.body3};
+  color: ${({ theme }) => theme.colors.grey900};
+  ${({ subText }) =>
+    subText &&
+    css`
+      color: ${({ theme }) => theme.colors.grey400};
+    `}
+  ${({ bold }) =>
+    bold &&
+    css`
+      font-weight: 600;
+    `}
+  margin-right: 5px;
+  &:last-child {
+    margin-right: 0;
+  }
 `;
