+++ packages/gdsc-dju-blog/src/components/molecules/WidthPostsContainer/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import PostCard from '@src/components/molecules/PostCard';
-import { PostCardWrapper } from './styled';
+import WidthPostCard from '@src/components/molecules/WidthPostCard';
+import { PostCardWrapper, WidthPostsContainerInner } from './styled';
 import { DetailPostDataType } from '@type/postData';
 import { useGetScrap } from '@src/hooks/useGetScrap';
 
@@ -11,10 +11,10 @@ interface Props {
 const WidthPostsContainer = ({ postData }: Props) => {
   const { scrapList } = useGetScrap();
   return (
-    <div>
+    <WidthPostsContainerInner>
       {postData.map((data) => (
         <PostCardWrapper key={data.postId}>
-          <PostCard
+          <WidthPostCard
             postData={data}
             isScrap={
               scrapList ? !!scrapList.find((id) => id == data.postId) : false
@@ -22,7 +22,7 @@ const WidthPostsContainer = ({ postData }: Props) => {
           />
         </PostCardWrapper>
       ))}
-    </div>
+    </WidthPostsContainerInner>
   );
 };
