+++ packages/gdsc-dju-blog/src/components/molecules/WidthPostsContainer/styled.ts
@@ -4,4 +4,8 @@ export const PostCardWrapper = styled.div`
   display: flex;
   flex-direction: row;
   margin-bottom: 30px;
+  width: 100%;
+`;
+export const WidthPostsContainerInner = styled.div`
+  width: 100%;
 `;
