+++ packages/gdsc-dju-blog/src/components/organisms/WidthPostsContainerWithMenu/index.tsx
@@ -1,36 +1,42 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import WidthPostsContainer from '@src/components/molecules/WidthPostsContainer';
 import PageBar from '@src/components/atoms/PageBar';
-import { DetailPostDataType } from '@type/postData';
+import { PostListResponse } from '@type/postData';
 import {
   CategoryContainer,
   PageBarWrapper,
 } from '@src/components/organisms/PostsContainerWithMenu/styled';
+import Notice from '@src/components/atoms/Notice';
 
 interface Props {
-  postData: DetailPostDataType[];
   currentPage: number;
-  totalPage: number;
-  pageHandler: (page: number, limit?: number) => void;
+  postListResponse: PostListResponse | undefined;
 }
 
 const WidthPostsContainerWithMenu = ({
-  postData,
+  postListResponse,
   currentPage,
-  totalPage,
-  pageHandler,
 }: Props) => {
   return (
-    <CategoryContainer>
-      <WidthPostsContainer postData={postData} />
-      <PageBarWrapper>
-        <PageBar
-          currentPage={currentPage}
-          totalPage={totalPage || 0}
-          onClick={pageHandler}
-        />
-      </PageBarWrapper>
-    </CategoryContainer>
+    <Suspense fallback={<div>postList isLoading</div>}>
+      {postListResponse && (
+        <CategoryContainer>
+          {postListResponse.empty ? (
+            <Notice>작성된 글이 없어요!</Notice>
+          ) : (
+            <>
+              <WidthPostsContainer postData={postListResponse.content} />
+              <PageBarWrapper>
+                <PageBar
+                  currentPage={currentPage}
+                  totalPage={postListResponse.totalPages || 0}
+                />
+              </PageBarWrapper>
+            </>
+          )}
+        </CategoryContainer>
+      )}
+    </Suspense>
   );
 };
