+++ packages/gdsc-dju-blog/src/pages/MyBlog/BlogHome/MyBlogContentBox/styled.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 
-export const MyBlogContentWrapper = styled.main`
+export const MyBlogContentWrapper = styled.div`
   width: 100%;
-  height: 600px;
 `;
