+++ packages/gdsc-dju-blog/src/pages/PostSaves/PostSavesMenuBox/styled.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const PostSavesMenuBoxWrapper = styled.menu`
+  display: flex;
+  align-items: flex-end;
+  flex-direction: row;
+  width: 100%;
+  justify-content: space-between;
+`;
