+++ packages/gdsc-dju-blog/src/pages/PostSaves/index.tsx
@@ -13,6 +13,7 @@ import { useGetMyData } from '@src/api/hooks/useGetMyData';
 import { ContainerInner, LayoutContainer } from '@styles/layouts';
 
 import WidthPostsContainerWithMenu from '@src/components/organisms/WidthPostsContainerWithMenu';
+import PostSavesLayout from '@src/components/layouts/PostSavesLayout';
 
 const PostSaves = () => {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -21,67 +22,25 @@ const PostSaves = () => {
   const pageParams = searchParams.get('page');
   const page = pageParams ? parseInt(pageParams) : 1;
 
-  const { userPostTempData } = useGetMyPostsTempData(category, page - 1, 6);
-  const navigate = useNavigate();
-
   useEffect(() => {
-    if (page) {
+    if (!pageParams && !categoryName) {
       setSearchParams({
         type: category,
         page: '1',
       });
     }
   }, []);
-
-  const pageHandler = (page: number, limit?: number) => {
-    if (page < 1) {
-      return;
-    }
-    if (page === limit) {
-      return;
-    } else {
-      navigate({
-        pathname: `/post/saves`,
-        search: `?${createSearchParams({
-          type: category,
-          page: page.toString(),
-        })}`,
-      });
-    }
-  };
   const categoryHandler = (category: string) =>
-    navigate({
-      pathname: `/post/saves`,
-      search: `?${createSearchParams({
-        type: category,
-        page: page.toString(),
-      })}`,
-    });
+    setSearchParams({ ...searchParams, type: category });
   return (
     <>
       <LayoutContainer>
         <ContainerInner>
-          <PostSavesTitle>임시 저장된 글</PostSavesTitle>
-
-          <CategoryMenu type={category as string} onClick={categoryHandler} />
-          <ButtonWrapper>
-            <GDSCButton
-              text={'새로작성'}
-              color={'googleBlue'}
-              onClick={() => navigate('/post/write')}
-            />
-          </ButtonWrapper>
-
-          <Suspense fallback={<div>isLoading</div>}>
-            {userPostTempData && (
-              <WidthPostsContainerWithMenu
-                postData={userPostTempData.content}
-                pageHandler={pageHandler}
-                currentPage={page}
-                totalPage={userPostTempData.totalPages}
-              />
-            )}
-          </Suspense>
+          <PostSavesLayout
+            page={page}
+            category={category}
+            categoryHandler={categoryHandler}
+          />
         </ContainerInner>
       </LayoutContainer>
     </>
