+++ packages/gdsc-dju-blog/src/pages/PostSaves/styled.ts
@@ -11,5 +11,4 @@ export const PostSavesTitle = styled.div`
   display: flex;
   font-size: ${(props) => props.theme.fontSize.h4};
   font-weight: 700;
-  margin: 110px 0 86px 0;
 `;
